+++ src/tests.rs
@@ -3,6 +3,8 @@
 mod tests {
     // Note this useful idiom: importing names from outer (for mod tests) scope.
     use native_regex_lib::native_regex::{NativeRegex, NativeRegexSet};
+    use native_regex_lib::native_regex::Captures;
+    use native_regex_lib::native_regex::NoExpand;
 
     include!(concat!(env!("OUT_DIR"), "/regexes.rs"));
 
@@ -16,17 +18,35 @@ mod tests {
         //Make sure the regex finds all the matches
         {
             let mut capture_iter = reg_test.captures_iter(haystack);
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "Qoij345");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "Ndds");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "Qoij345");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "Ndds");
         }
 
         //Make sure the regex correctly identifies the captures in the first regex
         {
             let first_match = reg_test.captures(haystack).unwrap();
-            assert_eq!(first_match.get(0).unwrap().as_str(), "Qoij345");
+            assert_eq!(first_match.first().as_str(), "Qoij345");
             assert_eq!(first_match.get(1).unwrap().as_str(), "Qoij");
             assert_eq!(first_match.get(2).unwrap().as_str(), "345");
         }
+    }
+
+    #[test]
+    fn optional_regex() {
+        let haystack = "hep";
+
+        let reg_test = OptionalRegex::new();
+
+        //Make sure the regex correctly identifies the captures in the first regex
+        {
+            let first_match = reg_test.captures(haystack).unwrap();
+            assert_eq!(first_match.first().as_str(), "hep");
+            assert_eq!(first_match.get(1).unwrap().as_str(), "h");
+            assert_eq!(first_match.get(2).unwrap().as_str(), "e");
+            assert_eq!(first_match.get(3), None);
+            assert_eq!(first_match.get(4).unwrap().as_str(), "p");
+        }
+
     }
 
 
@@ -72,12 +92,13 @@ mod tests {
         //Make sure the regex finds all the matches
         {
             let mut capture_iter = reg_test.captures_iter(float_search);
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "123.324e44");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "123");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "123e59");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "312.45");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "123.324e44");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "123");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "123e59");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "312.45");
         }
     }
+
 
     #[test]
     fn nonword_regex() {
@@ -88,15 +109,15 @@ mod tests {
         //Make sure the regex finds all the matches
         {
             let mut capture_iter = reg_test.captures_iter(nonword_search);
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ^*&%*&^ ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "-");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), "   ");
-            assert_eq!(capture_iter.next().unwrap().get(0).unwrap().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ^*&%*&^ ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "-");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), "   ");
+            assert_eq!(capture_iter.next().unwrap().first().as_str(), " ");
         }
 
     }
@@ -107,7 +128,7 @@ mod tests {
 
         let reg_test = GreyRegex::new();
 
-        assert_eq!(reg_test.replace(grey_search, |_, _| String::from("purple")).as_str(), "Text to show the replacing of purple and purple.");
+        assert_eq!(reg_test.replace(grey_search, |caps: &Captures| {"purple"}), "Text to show the replacing of purple and purple.");
 
     }
 
@@ -117,13 +138,14 @@ mod tests {
 
         let reg_test = Ipv4Regex::new();
 
-        assert_eq!(reg_test.replace(ip_search, |_, cap| {
+        assert_eq!(reg_test.replace(ip_search, |cap: &Captures| {
             let mut num = 0u32;
             let mut scale = 256*256*256u32;
+            let t = cap;
             for index in 1..5 {
                 match cap.get(index).unwrap().as_str().parse::<u8>() {
                     Ok(byte) => { num += byte as u32 * scale; scale /= 256; }
-                    Err(_) => return format!("{}", cap.get(0).unwrap().as_str())
+                    Err(_) => return format!("{}", cap.first().as_str())
                 }
             }
             format!("{}", num)
@@ -131,6 +153,7 @@ mod tests {
 
     }
 
+    /*
     #[test]
     fn wordboundary_regex() {
 
@@ -149,6 +172,9 @@ mod tests {
 
     }
 
+     */
+
+
     #[test]
     fn anchor_regex() {
 
@@ -160,7 +186,7 @@ mod tests {
         //Make sure the regex finds all the matches
         {
             let mut capture_iter = start_test.captures_iter(sample_string);
-            let m = capture_iter.next().unwrap().get(0).unwrap();
+            let m = capture_iter.next().unwrap().first();
             assert_eq!(m.as_str(), "hello");
             assert_eq!(m.start(), 0);
             assert_eq!(capture_iter.next(), None);
@@ -169,13 +195,47 @@ mod tests {
         //Make sure the regex finds all the matches
         {
             let mut capture_iter = end_test.captures_iter(sample_string);
-            let m = capture_iter.next().unwrap().get(0).unwrap();
+            let m = capture_iter.next().unwrap().first();
             assert_eq!(m.as_str(), "hello");
             assert_eq!(m.start(), 24);
             assert_eq!(capture_iter.next(), None);
         }
 
     }
+
+
+    #[test]
+    fn anchor_line_regex() {
+
+        let sample_string = "startolinesdokfpsdok \nstartofline endofline\n oaisjdoijendoline";
+
+        let start_test = StartLineRegex::new();
+        let end_test = EndLineRegex::new();
+
+        //Make sure the regex finds all the matches
+        {
+            let mut capture_iter = start_test.captures_iter(sample_string);
+            let m = capture_iter.next().unwrap().first();
+            assert_eq!(m.as_str(), "startoline");
+            let m = capture_iter.next().unwrap().first();
+            assert_eq!(m.as_str(), "startofline");
+            assert_eq!(capture_iter.next(), None);
+        }
+
+        //Make sure the regex finds all the matches
+        {
+            let mut capture_iter = end_test.captures_iter(sample_string);
+            let m = capture_iter.next().unwrap().first();
+            assert_eq!(m.as_str(), "endofline");
+            let m = capture_iter.next().unwrap().first();
+            assert_eq!(m.as_str(), "endoline");
+            assert_eq!(capture_iter.next(), None);
+        }
+
+    }
+
+
+
 
     #[test]
     fn name_test() {
@@ -189,6 +249,9 @@ mod tests {
         assert_eq!(cap.name("quantity").unwrap().as_str(), "33");
 
     }
+
+
+
 
     #[test]
     fn is_match_regset_test() {
@@ -208,6 +271,9 @@ mod tests {
 
     }
 
+
+
+
     #[test]
     fn matches_regset_test() {
         let ipv4 = Ipv4Regex::new();
@@ -216,12 +282,37 @@ mod tests {
 
         let set = NativeRegexSet::new(vec![ipv4.into(), symbol.into(), float.into()]);
 
-        assert_eq!(set.matches("testing!").iter().map(|x| *x).collect::<Vec<_>>(), Vec::<(usize, usize)>::new());
-        assert_eq!(set.matches("testTing!").iter().map(|x| *x).collect::<Vec<_>>(), vec![(1, 4)]);
-        assert_eq!(set.matches("192.168.0.0").iter().map(|x| *x).collect::<Vec<_>>(), vec![(0, 0), (2, 0)]);
-        assert_eq!(set.matches("sdfsd 192.168.0.0 He45 ad").iter().map(|x| *x).collect::<Vec<_>>(), vec![(0, 6), (2, 6), (1, 18)]);
+        assert_eq!(set.matches("testing!").iter().map(|(regex, capture)| (*regex, capture.first().start())).collect::<Vec<_>>(), Vec::<(usize, usize)>::new());
+        assert_eq!(set.matches("testTing!").iter().map(|(regex, capture)| (*regex, capture.first().start())).collect::<Vec<_>>(), vec![(1, 4)]);
+        assert_eq!(set.matches("192.168.0.0").iter().map(|(regex, capture)| (*regex, capture.first().start())).collect::<Vec<_>>(), vec![(0, 0), (2, 0)]);
+        assert_eq!(set.matches("sdfsd 192.168.0.0 He45 ad").iter().map(|(regex, capture)| (*regex, capture.first().start())).collect::<Vec<_>>(), vec![(0, 6), (2, 6), (1, 18)]);
 
+        let set_matches = set.matches("sdfsd 192.168.0.0 He45 ad");
+        let mut set_matches_iter = set_matches.iter().map(|(regex, capture)| (*regex, capture.first().as_str()));
 
+        assert_eq!(set_matches_iter.next(), Some((0, "192.168.0.0")));
+        assert_eq!(set_matches_iter.next(), Some((2, "192.168")));
+        assert_eq!(set_matches_iter.next(), Some((1, "He45")));
+
+    }
+
+    #[test]
+    fn empty_regex_test() {
+        let reg_test = EmptyRegex::new();
+
+        assert_eq!(reg_test.find("asitgdoisd").unwrap().as_str(), "tg");
+
+    }
+
+    #[test]
+    fn replacer_regex_test() {
+        let reg_test = SymbolRegex::new();
+
+        let haystack = "things like He2 and Na13 also H and F";
+
+        assert_eq!(reg_test.replace(haystack, "(S: ${1}, Q: ${2})"), format!("things like (S: He, Q: 2) and (S: Na, Q: 13) also (S: H, Q: ) and (S: F, Q: )"));
+        assert_eq!(reg_test.replace(haystack, NoExpand::new("CHEMICAL")), format!("things like CHEMICAL and CHEMICAL also CHEMICAL and CHEMICAL"));
+        assert_eq!(reg_test.replace(haystack, String::from("(${1}, ${2})")), format!("things like (He, 2) and (Na, 13) also (H, ) and (F, )"));
     }
 
 }